+++ packages/task-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @opentask/taskin-task-manager
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @opentask/taskin-types@1.0.6
+
 ## 1.0.8
 
 ### Patch Changes